requirements.txt
absl-py==2.1.0
aiofiles==23.2.1
aiogram==3.12.0
aiohappyeyeballs==2.3.6
aiohttp==3.10.3
aiosignal==1.3.1
aiosqlite==0.20.0
annotated-types==0.7.0
attrs==24.2.0
certifi==2024.7.4
cffi==1.17.0
click==8.1.7
contourpy==1.2.1
cycler==0.12.1
easyocr==1.7.1
filelock==3.15.4
flatbuffers==24.3.25
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.6.1
greenlet==3.0.3
idna==3.7
imageio==2.35.0
jax==0.4.31
jaxlib==0.4.31
Jinja2==3.1.4
kiwisolver==1.4.5
lazy_loader==0.4
magic-filter==1.0.12
MarkupSafe==2.1.5
matplotlib==3.9.2
mediapipe==0.10.14
ml-dtypes==0.4.0
mpmath==1.3.0
multidict==6.0.5
networkx==3.3
ninja==1.11.1.1
numpy==2.0.1
ocrmac==0.1.6
opencv-contrib-python==4.10.0.84
opencv-python-headless==4.10.0.84
opt-einsum==3.3.0
packaging==24.1
pillow==10.4.0
protobuf==4.25.4
pyclipper==1.3.0.post5
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pyobjc-core==10.3.1
pyobjc-framework-Cocoa==10.3.1
pyobjc-framework-CoreML==10.3.1
pyobjc-framework-Quartz==10.3.1
pyobjc-framework-Vision==10.3.1
pyparsing==3.1.2
python-bidi==0.6.0
python-dateutil==2.9.0.post0
PyYAML==6.0.2
scikit-image==0.24.0
scipy==1.14.0
shapely==2.0.5
six==1.16.0
sounddevice==0.5.0
SQLAlchemy==2.0.32
sympy==1.13.2
tifffile==2024.8.10
torch==2.4.0
torchvision==0.19.0
typing_extensions==4.12.2
yarl==1.9.4
